refactor: PR 리뷰 반영 - ProductListViewModel에서 cell을 위한 뷰모델 분리
